Refactor gatsby-config.ts
